+++ Changelog.md
@@ -36,6 +36,14 @@
 
 ### Ignition Physics 1.x.x (20XX-XX-XX)
 
+1. Fix collision issue with detachable joints
+    * [Pull request 31](https://github.com/ignitionrobotics/ign-physics/pull/31)
+
+1. Add PlaneShape feature and implement in dartsim with test.
+    * [BitBucket pull request 66](https://osrf-migration.github.io/ignition-gh-pages/#!/ignitionrobotics/ign-physics/pull-requests/66)
+
+### Ignition Physics 1.7.0 (2020-04-13)
+
 1. Add RequestFeatures API for casting the features of an entity to a new feature set when possible.
     * [BitBucket pull request 130](https://osrf-migration.github.io/ignition-gh-pages/#!/ignitionrobotics/ign-physics/pull-requests/130)